fix errors with dumb/unknown terminals
With dumb terminals (like the shells emacs spawns), the various ncurses
commands (like `tput`) will exit non-zero which will cascades down as the
last exit code in the script because it's the last thing we ran.

This breaks scripts that do something like:
	source /lib/gentoo/functions.sh || exit 1

If we make it to the end of the script, just clear the last exit status.