consoletype.c: update header to <sys/sysmacros.h>
This is an update to commit 53b60b4.  Recent changes in musl have
removed the inclusion of <sys/sysmacros.h> from <sys/types.h>.
Now the macros major()/minor() require the inclusion of the former
header.  This change does not affect glibc nor earlier versions of
musl, but is required for musl 1.1.23 and later.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>