Avoid unnecessary subshells and use hash to avoid repeated PATH lookups
Use hash to determine whether logger is available, preventing repeated
PATH searches in the case that esyslog() is called more than once.

Once RC_NOCOLOR is set to "yes", don't waste any further time iterating
over the remaining positional parameters.

Use hash to determine whether tput is available, preventing at least one
repeated PATH lookup.

Don't incur a pointless subshell in the course of checking whether tput
is available on the second occasion.

Don't repeatedly execute tput, incurring a subshell on each occasion, to
get the sequences for "bold" and "sgr0". Obtain both, once.

Use unset -v to indicate that variables, specifically, are to be unset.

Remove the true hack at the end of the script. It's no longer needed.

The overall reduction in the number of subshells amounts to 9 and it can
be trivially proven that the code is executed more quickly.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>