Don't unsuccessfully attempt to background the defunct rc_splash utility
The media-gfx/splashutils package was dropped in November 2018, and was
probably not even usable for some time before that. Currently, every
time the _eend() function handles an exit status code other than 0, it
tries to run rc_splash in the background with no possible chance of
success. It also incurs an utterly useless syscall, in order to
determine whether /dev/null is a character device. Cease and desist.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>