Re-implement get_bootparam as pure sh, eliminating the gawk dependency
The current implementation of get_bootparam() specifically invokes gawk
and has some deficiencies, such as failing to inhibit pathname
expansion and not breaking the loop once a match has been made.

In fact, the problem can be solved using sh(1) alone, and the revised
function does precisely that. Note that the definition of the function
is now the kind of compound command that incurs a subshell. Hence, one
subshell is incurred for the entire routine and there is no compelling
reason to use the (non-standard) local keyword, nor be concerned with
the consequences of disabling pathname expansion.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Bug: https://bugs.gentoo.org/886017
Signed-off-by: Sam James <sam@gentoo.org>