Don't unconditionally shift 2 in esyslog()
Where esyslog() decides that it is appropriate to log a message, it
unconditionally tries to shift two of of the positional parameters,
regardless of how many were given. As has been seen before, this will
cause a non-interactive instance of sh(1) to immediately exit.

The other functions that call into it are not able to trigger this bug.
Still, it's possible for esyslog() to be directly invoked with too few
arguments. Fix it by testing whether a minimum of two arguments were
given. Should the test fail, print an error message and return 1.

Also, improve the method by which superfluous messages are ignored. The
prior strategy was to join the messages by the first character of IFS,
then check that the length of the resulting string is greater than 0.
Instead, check that the resulting string contains at least one visible
character.

Finally, mention that the invocation of logger(1) is not portable.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>