Consistently specify the tag where calling esyslog()
In some cases, the tag name is specified as "rc-scripts" whereas, in
others, it is specified as "${0##*/}". Consistently use the latter.

Eliminate some useless uses of the (non-portable) local builtin.

Convey the positional parameters to esyslog() with "$@". They end up
being re-joined, so they may as well be conveyed exactly as they were.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>