Improve the means by which the number of terminal columns is detected
The existing routine is not well structured and is perhaps a little too
trusting of its inputs. Clean it up and check that the value of COLUMNS
and the second field obtained from stty(1) are actually integers. To
this end, an is_int() function has been added. This function may yet
prove to be useful elsewhere.

This also fixes an instance of SC2068 by using read to consume the
output of stty(1), rather than perform word splitting and/or pathname
expansion. It also fixes an instance of SC2059 (format string abuse).

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>