Ensure that COLUMNS is set in the case that /bin/sh is bash
To reasonably ensure that the COLUMNS variable is set requires for
three conditions to be fulfilled. Firstly, the version of bash must be
at least 4.3. Secondly, the checkwinsize shopt must be enabled (since
5.0, it is enabled by default). Thirdly, an external command must be
executed.

Better fulfil these conditions by checking whether the shell appears to
be bash and, should it appear to be, enabling checkwinsize before
executing /bin/true. Determining the number of columns in this way is
preferable to falling back to the method of attempting to use stty(1).

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>