Add an _is_visible() function and use it internally
Joining one or more positional parameters by the first character of IFS
then checking whether the resulting string has a length of zero, is not
an effective way of determining whether the collective message is
blank. This implements a helper function that implements the check in
the fashion that was seemingly intended.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>