Improve the __end(), eend() and ewend() functions
The _eend() function has a poor calling convention because, while both
eend() and ewend() always pass a function name, _eend() expects for it
to be the second argument. A better design would be for mandatory
arguments to precede optional ones. Make it so.

Also, implement a parameter validation routine for the optionally
specified exit status code. It determines whether the parameter is in
the form of an integer and whether its value is greater than or equal
to 0. Should the parameter be considered invalid, a diagnostic message
will be printed to STDERR and its value will be taken as if it were 0.
Note that negative values are not necessarily tolerated by builtins such
as return and exit. As is shown below, they may cause sh(1) to exit.

$ dash -c 'return -1; echo done'
dash: 1: return: Illegal number: -1

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>