Validate the exit status given to veend() and vewend() as an argument
Presently, both functions trust that the first argument is either empty
or a valid exit status code. If given a bogus value, the call to return
may cause sh(1) to unceremoniously exit, as demonstrated below.

$ dash -c 'return junk; echo done'
dash: 1: return: Illegal number: junk

Instead, validate that the argument is an integer whose value is greater
than or equal to 0. Should validation fail, a diagnostic mesage will be
printed to STDERR and the return value will be 0, provided that the call
to print succeeded.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>