Implement - and integrate - an _eprint() helper function
The _eprint() function encapsulates some of the behaviour that the
following functions have in common.

  - ebegin
  - einfo
  - einfon
  - eerror
  - eerrorn
  - ewarnn
  - ewarn

Additionally, have einfo() wrap einfon(), eerror() wrap eerrorn(), and
ewarn() wrap ewarnn().

All of this allows for a pleasing degree of code reduction while
improving the structure of the functions overall, partly owing to a
complementary reduction in the number of return statements. It also
plots a course for the eventual implementation of proper TTY detection,
in a context-sensitive manner.

Bug: https://bugs.gentoo.org/730432
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>