Ensure that ebegin() appends a newline to the final message
The ebegin() function was recently modified to strip the trailing <newline>,
if any, from the given message. Later, it was modified to strip all trailing
<newline> characters. The purpose of doing so is to ensure that there is
no accidental line breaking betwixt the message and the ASCII ellipsis
that is appended. However, the function should then have been appending a
single <newline> to the composed message before handing it off to _eprint().
Make it so.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Fixes: 1947f0ed81f3b95a7e10a8a5a707776948f8a487