Fix the order of the arguments used to compose the CUP sequence
The CUP (ECMA-48 CSI) sequence expects for the row to come first, and
the column second. I could have just swapped the arguments but, instead,
I have adjusted the surrounding code so as to strictly adhere to this
convention (one that is also observed by stty size). This should
eliminate the possibility of any such mistake being made in the future.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Fixes: 20bc15b5b1009149c4a3d531911d3c219dc55f3a